Edit templates in memory, save to db only on confirm
Fixes #5011

Note this does not include undo functionality - it was difficult
enough just trying to split this out into reviewable chunks. That is
not a regression though, AnkiDroid had "discard changes" but not undo,
and it only really worked for template content. If you discard changes
here they really all go away.